+++ project/src/components/offer-card/offer-card.tsx
@@ -1,8 +1,8 @@
 import { useAppDispatch } from '../../hooks';
 import { setActiveOffer } from '../../store/action';
 import { generatePath, Link } from 'react-router-dom';
-import { Offer } from '../../types/data';
-import { StyleOfferCard } from '../../types/common';
+import { Offer } from '../../types/common';
+import { StyleOfferCard } from '../../types/properties-style';
 import { AppRoute } from '../../const';
 
 type OfferCardProps = {
