+++ project/src/store/action.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@reduxjs/toolkit';
-import { Offer, Offers, Reviews } from '../types/data';
+import { Offer, Offers, Review, Reviews } from '../types/common';
 import { AppRoute, AuthorizationStatus } from '../const';
 
 export const changeCity = createAction<{city: string}>('offers/changeCity');
@@ -19,6 +19,7 @@ export const setOffersLoadingStatus = createAction<boolean>('offers/setOffersLoa
 export const setSelectedOfferLoadingStatus = createAction<boolean>('offer/setSelectedOfferLoadingStatus');
 export const setNearbyOffersLoadingStatus = createAction<boolean>('offer/setNearbyOffersLoadingStatus');
 export const setReviewsLoadinStatus = createAction<boolean>('offer/setReviewsLoadinStatus');
+export const setFormReviewData = createAction<Review>('offer/setFormReviewData');
 
 export const requireAuthorization = createAction<AuthorizationStatus>('user/requireAuthorization');
 export const getUserEmail = createAction<string>('user/getUserEmail');
