+++ project/src/types/properties-style.ts
@@ -1,0 +1,9 @@
+export type PropertiesMap = {
+  className: string;
+  height: string;
+}
+
+export type StyleOfferCard = {
+  classArticle: string;
+  classImageWrapper: string;
+};
